+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Add validation to require PURLs (OSIDB-4810)
+- Automatic affectedness and resolution determination based on product
+  definitions (OSIDB-4938)
 
 ## [5.10.3] - 2026-05-06
 ### Fixed